Smashed old-style resistances/immunities/vulnerabilities. They are automatically converted to %% in load_race, load_mobiles, load_old_mobile, load_objects. Dual wield skill for mobiles.
git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@1911 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/include/race.h
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: race.h,v 1.18 1999-12-02 10:54:09 kostik Exp $
+ * $Id: race.h,v 1.19 1999-12-10 11:30:06 kostik Exp $
  */
 
 #ifndef _RACE_H_
@@ -44,7 +44,7 @@ struct race_t
 	flag32_t	form;		/* default form flag		*/
 	flag32_t	parts;		/* default body parts		*/
 	flag32_t	race_flags;	/* race flags			*/
-	int		resists[MAX_RESIST];
+	int16_t		resists[MAX_RESIST];
 					/* initial resistances		*/
 	pcrace_t *	race_pcdata;	/* additional data for pc races */
 };
